Refactor db code using Abstract Factory pattern
[FEATURE]

- The db factory will allow us to vend different db objects
- This provides a better structure to handle multiple db tables
- Moved parse-specific classes into database/parse directory
- Added unit tests

[TESTING DONE]

- Ran tslint
- Ran yarn test
- Ran npm expo

[RESOURCES]

- jira: https://peacefulli.atlassian.net/browse/PFML-22
