--- src/database/models/event-model.ts
@@ -1,11 +1,0 @@
-export default interface EventModel {
-  title?: string;
-  description?: string;
-  location?: string;
-
-  // URL for the image representing the event
-  imageSource?: string;
-
-  // JavaScript `Date` object representing when the event begins
-  dateTime?: Date;
-}
