+++ tst/App.spec.tsx
@@ -1,10 +1,10 @@
-import { default as App, default as Database } from "../App";
+import { default as App, default as Databases } from "../App";
 
-jest.mock("../src/database/parse-server");
+jest.mock("../src/database/parse/parse-database-factory");
 
 describe("Main App Tests", () => {
   it("exports are non-null", async () => {
-    expect(Database).not.toBeNull();
+    expect(Databases).not.toBeNull();
     expect(App).not.toBeNull();
   });
 });
